+++ src/ome_zarr_models/v05/image.py
@@ -1,5 +1,6 @@
 from pydantic import Field
 
+from ome_zarr_models.common.omero import Omero
 from ome_zarr_models.v05.base import BaseGroupv05, BaseOMEAttrs
 from ome_zarr_models.v05.multiscales import Multiscale
 
@@ -8,7 +9,7 @@ __all__ = ["Image", "ImageAttrs"]
 
 class ImageAttrs(BaseOMEAttrs):
     """
-    Model for the metadata of OME-Zarr data.
+    Metadata for OME-Zarr image groups.
     """
 
     multiscales: list[Multiscale] = Field(
@@ -16,6 +17,7 @@ class ImageAttrs(BaseOMEAttrs):
         description="The multiscale datasets for this image",
         min_length=1,
     )
+    omero: Omero | None = None
 
 
 class Image(BaseGroupv05[ImageAttrs]):