+++ tests/v04/test_hcs.py
@@ -2,12 +2,12 @@ from pathlib import Path
 
 import zarr
 
+from ome_zarr_models.common.omero import Channel, Omero, Window
 from ome_zarr_models.v04.axes import Axis
 from ome_zarr_models.v04.coordinate_transformations import VectorScale
 from ome_zarr_models.v04.hcs import HCS, HCSAttrs
 from ome_zarr_models.v04.image import ImageAttrs
 from ome_zarr_models.v04.multiscales import Dataset, Multiscale
-from ome_zarr_models.v04.omero import Channel, Omero, Window
 from ome_zarr_models.v04.plate import Acquisition, Column, Plate, Row, WellInPlate
 from ome_zarr_models.v04.well_types import WellImage, WellMeta